refactoring/rearranging code
